+++ src/SparseLUSolver.cpp
@@ -6,57 +6,48 @@
 // SPDX-License-Identifier: MPL-2.0
 // This file is part of LightSim2grid, LightSim2grid implements a c++ backend targeting the Grid2Op platform.
 
-#include "KLUSolver.h"
+#include "SparseLUSolver.h"
 
-void KLUSolver::reset(){
-    BaseNRSolver::reset();
-    klu_free_symbolic(&symbolic_, &common_);
-    klu_free_numeric(&numeric_, &common_);
-    n_ = -1;
-    common_ = klu_common();
-
-    symbolic_ = nullptr;
-    numeric_ = nullptr;
-}
-
-void KLUSolver::initialize(){
+void SparseLUSolver::initialize(){
     // default Eigen representation: column major, which is good for klu !
-    // J is const here, even if it's not said in klu_analyze
+    // J is const here
     auto timer = CustTimer();
     n_ = J_.cols(); // should be equal to J_.nrows()
     err_ = 0; // reset error message
-    common_ = klu_common();
-    symbolic_ = klu_analyze(n_, J_.outerIndexPtr(), J_.innerIndexPtr(), &common_);
-    numeric_ = klu_factor(J_.outerIndexPtr(), J_.innerIndexPtr(), J_.valuePtr(), symbolic_, &common_);
-    if (common_.status != KLU_OK) {
+    J_.makeCompressed();
+    solver_.analyzePattern(J_);  //NEW
+    solver_.factorize(J_);  // NEW
+    if (solver_.info() != Eigen::Success) {
         err_ = 1;
     }
     need_factorize_ = false;
     timer_solve_ += timer.duration();
 }
 
-void KLUSolver::solve(Eigen::VectorXd & b, bool has_just_been_inialized){
+void SparseLUSolver::solve(Eigen::VectorXd & b, bool has_just_been_inialized){
+    // NEW
+
     // solves (for x) the linear system J.x = b
-    // supposes that the solver has been initialized (call klu_solver.analyze() before calling that)
+    // supposes that the solver has been initialized (call sparselu_solver.analyze() before calling that)
     // J is const even if it does not compile if said const
     auto timer = CustTimer();
-    int ok;
     bool stop = false;
     if(!has_just_been_inialized){
         // if the call to "klu_factor" has been made this iteration, there is no need
         // to re factor again the matrix
         // i'm in the case where it has not
-        ok = klu_refactor(J_.outerIndexPtr(), J_.innerIndexPtr(), J_.valuePtr(), symbolic_, numeric_, &common_);
-        if (ok != 1) {
+        solver_.factorize(J_);  // NEW
+        if (solver_.info() != Eigen::Success) {
             err_ = 2;
             stop = true;
         }
     }
     if(!stop){
-        ok = klu_solve(symbolic_, numeric_, n_, 1, &b(0), &common_);
-        if (ok != 1) {
+        Eigen::VectorXd Va = solver_.solve(b);  //NEW
+        if (solver_.info() != Eigen::Success) {
             err_ = 3;
         }
+        b = Va;  // NEW
     }
     timer_solve_ += timer.duration();
 }